webhook_receiver/requirements.txt
azure-identity
azure-keyvault-secrets
fastapi
redis
requests
loguru
urllib3
pydantic
deepdiff
uvicorn[standard]